Complete of rebuilding the integration tests
