+++ test/MoneyManager.API.IntegrationTests/ControllerTests/TrackerControllerTests/DeleteRecordTests.cs
@@ -8,6 +8,7 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 using Xunit;
+using MoneyManager.API.IntegrationTests.ControllerTestUtils;
 
 namespace MoneyManager.API.IntegrationTests.ControllerTests.TrackerControllerTests
 {
